serve/restic: factor object serving into cmd/httplib/serve
